+++ packages/ui/src/components/avatar/index.ts
@@ -1,1 +1,17 @@
-export { Avatar } from './Avatar';
+export { Avatar, useAvatar, useAvatarContext } from './Avatar';
+export type {
+  AvatarContextProps,
+  AvatarFallbackBaseProps,
+  AvatarFallbackProps,
+  AvatarImageBaseProps,
+  AvatarImageProps,
+  AvatarRootBaseProps,
+  AvatarRootProps,
+  AvatarRootProviderBaseProps,
+  AvatarRootProviderProps,
+  AvatarSize,
+  AvatarStatusChangeDetails,
+  UseAvatarContext,
+  UseAvatarProps,
+  UseAvatarReturn,
+} from './Avatar';